Added whitelist for vulture
vulture supports a whitelist module that can be used to ignore certain
functions, as well as class attributes. Thus, whitelist was added to
ignore the assignment of certain class attributes (i.e. bokeh figure
assignments of attributes like labels and colors), as well as routes
for the Flask app (and assignment of attributes).

Also renamed a variable in db.py, as vulture also supports ignoring
variables that start with an underscore.
	modified:   density/db/db.py
	new file:   whitelist.py