Improved API key generation
Previous implementation simply used os.random to produce 32 random
characters; urandom > random as it isn't blocking (on Linux this is
only a risk when initially installing the OS). Also appended a
randomly generated UUID to increase guarantee of uniqueness.

Also added Base64 encoding to token as per standard protocol for
Authorization tokens.

Changes to be committed:
	modified:   density/db/db.py