Added special error message in case of key collision
In the very unlikely chance that a newly generated
API key results in a collision with an already existing
key in the database, there is currently no error handling
except for a failure message on the page that says
"An error occurred. Please try again."

This is the same error message displayed for literally any
exception that is thrown when attempting to authenticate,
so added specific handling logic with error message for the
case of a key collision; this will throw an IntegrityError,
which only occurs when an action threatens the integrity of
a database (e.g. violates a key check).
	modified:   density/density.py